feat(tron): add misc features (#160)
* feat(tron): add support for tron sign-messageV2 signing

1. add lock_period support for delegate-resource contract

* feat(legacy): add support for EIP-7702

* feat(legacy): optimize the display of tron energy rental

* feat(legacy): add support for Aptos SIWA

* chore(legacy): bump version to 3.13.0

+++ legacy/util.c
@@ -118,7 +118,7 @@ bool bracket_replace(char *orig, const char *with) {
       }
       p++;
     } else {
-      steps = utf8_get_size(*p);
+      steps = utf8_get_size((uint8_t)*p);
       p += steps;
     }
   }
